Modifying Unit Loading
Previously units were loaded into the pkg.env.
This is problematic as some more bespoke users
will want access to the unit tablesz, either for
writing them to file, for inspections, or to
modify units, and the pkg.env file was not
as accessible.

I have now loaded the units into a "unit"
object which can be accessed but is not
assigned to the global env: